+++ extensions/ql-vscode/src/data-extensions-editor/shared/hide-modeled-apis.ts
